fix: resolve mcp-server-os-level bugs and eliminate clippy warnings
Bug fixes (14 files):
- transmute → safe CGEventType::ScrollWheel constant
- panic!() → Err(AutomationError::Internal) in find_element/find_elements
- from_utf8_unchecked → String::from_utf8_lossy (3 sites)
- String byte-slicing → strip_prefix() in selector.rs
- Role case mismatch: match arms now lowercase to match input
- Locator::first() returns Ok(None) on ElementNotFound
- .unwrap() on chars().next() → safe alternatives
- unsafe Send+Sync documentation expanded
- Clippy: 25→0 warnings across all files

+++ src-tauri/mcp-server-os-level/src/platforms/macos/attributes.rs
@@ -126,7 +126,7 @@ pub(crate) fn parse_ax_attribute_value(
                     // Create an array of element addresses
                     let mut items = Vec::with_capacity(count as usize);
                     for i in 0..count {
-                        let item = CFArrayGetValueAtIndex(array_ref, i as isize);
+                        let item = CFArrayGetValueAtIndex(array_ref, i);
                         if !item.is_null() {
                             // Correctly wrap the raw pointer into AXUIElement
                             let ax_element = AXUIElement::wrap_under_get_rule(item as *mut _);
